Fix 50 build errors across 14 files
WarningsManager.cs:
- Add using Newtonsoft.Json.Linq for JObject/JArray/JToken
- Fix WarningReport.Critical → BySeverity.GetValueOrDefault()
- Fix SpatialElementTag → IndependentTag (correct Revit API type)
- Fix tuple .Priority access → .Item4 positional access
- Fully qualify ComplianceScan.ComplianceResult in external classes

WorkflowEngine.cs:
- Fix namespace for 5 commands (Temp/Tags → BIMManager)
- Comment out 2 unimplemented command references
- Remove duplicate "AutoAssignTemplates" switch case

ValidateTagsCommand.cs:
- Fix completePlaceholder/incomplete → correct variable names

ModelCommands.cs:
- Fix AutoTagAndReport call argument count

ModelEngine.cs:
- Fix RunFullPipeline seqCounters parameter type

ComplianceScan.cs:
- Fix RevitLinkType.GetLinkedDocument → linkInst.GetLinkDocument()

PlasteringEngine.cs:
- Fix MaterialFunctionAssignment.Finish → .None

BIMManagerCommands.cs:
- Fix tuple Direction/DeltaPct → Item1/Item2 access

StructuralAnalysisEngine.cs:
- Fix Math.Ceiling ambiguity with explicit double cast

StructuralModelingCommands.cs:
- Fix missing extension parameter and SaveSeqSidecar args

ExcelStructuralEngine.cs:
- Fix SaveSeqSidecar missing parameter

TokenWriterCommands.cs:
- Fix CheckComplianceGate missing commandName parameter

DataPipelineCommands.cs:
- Fix RevitLinkType.GetLinkedDocument → linkInst.GetLinkDocument()

StructuralIntelligenceEngine.cs:
- Fix XYZ.HasValue/Value → direct XYZ access (not nullable)

+++ StingTools/Model/StructuralAnalysisEngine.cs
@@ -1525,7 +1525,7 @@ namespace StingTools.Model
                         PhaseName = $"{level.Name} — Floor Slab",
                         Description = $"Cast {floors.Count} floor slabs",
                         ElementIds = floors.Select(e => e.Id).ToList(),
-                        EstimatedDays = Math.Max(2, (int)Math.Ceiling(floors.Count * 2)),
+                        EstimatedDays = Math.Max(2, (int)Math.Ceiling((double)(floors.Count * 2))),
                         LevelName = level.Name,
                     });
                 }
